requirements.txt
keras~=2.9.0
numpy~=1.23.0
tensorflow~=2.9.1
tf-models-official~=2.9.2
tqdm~=4.64.0
yacs~=0.1.8
wandb~=0.13.1
scipy~=1.8.1
matplotlib~=3.5.2
contextlib2~=21.6.0
tensorflow-io